use int instead of forced 8-bit types where appropriate

+++ shared/src/OctalCode.cpp
@@ -108,7 +108,7 @@ float * firstVertexForCode(unsigned char * octalCode) {
     memset(firstVertex, 0, 3 * sizeof(float));
     
     for (int i = 0; i < numberOfThreeBitSectionsInCode(octalCode); i++) {
-        int8_t sectionIndex = sectionValue(octalCode + 1 + (3 * i / 8), (3 * i) % 8);
+        int sectionIndex = sectionValue(octalCode + 1 + (3 * i / 8), (3 * i) % 8);
         for (int j = 0; j < 3; j++) {
             firstVertex[j] += 0.5 * (int)oneAtBit(sectionIndex, 7 -j);
         }
