+++ shared/src/SharedUtil.h
@@ -22,11 +22,11 @@ double usecTimestamp(timeval *time);
 double usecTimestampNow();
 
 float randFloat();
-unsigned char randomColorValue(uint8_t minimum);
+unsigned char randomColorValue(int minimum);
 bool randomBoolean();
 
 void outputBits(unsigned char byte);
-int8_t numberOfOnes(unsigned char byte);
-bool oneAtBit(unsigned char byte, int8_t bitIndex);
+int numberOfOnes(unsigned char byte);
+bool oneAtBit(unsigned char byte, int bitIndex);
 
 #endif /* defined(__hifi__SharedUtil__) */
